WG4: the shell — view bar, lazy views, docked transcript, live metrics
docs/ROADMAP.md's WG4 row, on WG3's substrate and in its grammar. The Mac's
64_cocoaui.mst is the reference design and this is the same grammar in
Windows-11 materials: a VIEW BAR across the top (switchers packed left, Do
It / Print It, metrics cluster anchored right), a CONTENT HOST filling the
middle with one child per view, and a collapsible TRANSCRIPT docked at the
bottom, newest line first.

world/93_winui_shell.mst — the layer, overriding 92's hooks:
  * WinLayout gains BOTTOM-anchored bands and `packRow:left:right:gap:dpi:`.
    Both stay pure arithmetic, so both are testable with no window; the
    bottom band is an ordinary `bands` entry carrying a third element
    (`#bottom`) because a class's shape cannot change once it exists.
  * A VIEW REGISTRY with genuinely lazy build: a view's content child is
    created on first show and never again (measured live: 3 views
    registered, 1 built at open, 2 after one switch, still 2 after
    switching to the same view again, 3 after visiting the third).
  * The dock collapses by SETTING ITS BAND'S HEIGHT to 0 rather than
    removing the band, so the collapsed layout is the same arithmetic as
    the open one and reopening restores a number, not a structure.
  * `updateMetricsMem:jit:code:alloc:gc:` — the Mac's own selector, one
    call because five values belong to one sample.
  * Do It / Print It enablement tracked by the active view (today: false,
    honestly, until WG5's Workspace can supply source).

world/92_winui_controls.mst — its three hook implementations gain wg3-
prefixed names and the hooks become one-line delegations. `super` cannot
serve a later layer here: reopening a class creates no superclass link to
its own previous definition, so a hook that could only be replaced
wholesale would force every future sprint to copy this one.

win_gui/src/main.rs + justfile — `MACVM_WINUI_WG4=off` restores WG3's
window exactly, and gate-wg3 now pins it, for the reason WG3 pinned
MACVM_WINUI_CONTROLS: a gate that can only run the newest layer cannot
bisect a regression to a sprint.

world/tests/64_winui_shell_tests.mst — ten headless tests over the layout
grammar and the registry, including one pinning WG3's own layout unchanged.
World suite: 7854 run, 0 failed.

Driven live through MACVM_WINUI_CTL and captured: target/winui-wg4b.png.
Two defects found by the first snap and fixed: a Win32 multiline EDIT
breaks on CRLF and on nothing else (a bare CR ran every transcript line
together), and WG3's demo trio, which has no band in this grammar, is
hidden rather than left drawing where it was created.

+++ gui/src/shell/win.rs
@@ -280,7 +280,7 @@ extern "system" fn wndproc(window: HWND, msg: u32, w: WPARAM, l: LPARAM) -> LRES
                 LRESULT(0)
             }
             WM_COMMAND => {
-                dispatch_menu_command((w.0 & 0xffff) as usize);
+                dispatch_menu_command(w.0 & 0xffff);
                 LRESULT(0)
             }
             // Keep the web view filling the client area.
